Add test_api
